% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_artistic.R
\name{artistic_theme}
\alias{artistic_theme}
\title{remove all axis, background and legend}
\usage{
artistic_theme(b_background = TRUE)
}
\arguments{
\item{b_background}{boolean}
}
\value{
theme
}
\description{
remove all axis, background and legend
}
