% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_shortcut.R
\name{help_shortcut}
\alias{help_shortcut}
\title{function to display help}
\usage{
help_shortcut()
}
\value{
automatic call of help on the function where the pointer is
}
\description{
function to display help
}
