% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lazy_variable_allocation_function.R
\name{set_lazy_variable_allocation_function}
\alias{set_lazy_variable_allocation_function}
\title{automatic writting of begining of allocation to  a variable inside rstudio}
\usage{
set_lazy_variable_allocation_function()
}
\value{
automatic writting of begining of allocation to  a variable inside rstudio
}
\description{
automatic writting of begining of allocation to  a variable inside rstudio
}
