% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_artistic.R
\name{theme_artistic}
\alias{theme_artistic}
\title{remove all axis, background and legend}
\usage{
theme_artistic(b_background = TRUE)
}
\arguments{
\item{b_background}{boolean}
}
\value{
theme
}
\description{
remove all axis, background and legend
}
